# Add optimised HyperLogLog++ implementation