This PR adds a custom HLL++ implementation.

It's adapted (a lot is plain copied) from an implementation provided by Clark DuVall https://github.com/clarkduvall/hyperloglog.

The differences are that this implementation:

 * uses an [AMD64 optimised xxhash](https://github.com/cespare/xxhash) algorithm instead of murmur3;
 * uses some AMD64 optimisations for things like [clz](https://en.wikipedia.org/wiki/Find_first_set#CLZ);
 * works with []byte rather than a Hash64 interface, to reduce allocations; 
 * implements encoding.BinaryMarshaler and encoding.BinaryUnmarshaler

Based on some rough benchmarking, this implementation of HyperLogLog++ is around twice as fast as the https://github.com/clarkduvall/hyperloglog implementation.

In terms of benchmarking how quickly values can be added to the HLL here are some numbers taken from an i5 MBP:

```
BenchmarkPlus_Add_100-4         	  300000	      5580 ns/op
BenchmarkPlus_Add_1000-4        	    5000	    361586 ns/op
BenchmarkPlus_Add_10000-4       	     200	   9886825 ns/op
BenchmarkPlus_Add_100000-4      	     500	   2073753 ns/op
BenchmarkPlus_Add_1000000-4     	     100	  20063534 ns/op
BenchmarkPlus_Add_10000000-4    	       5	 204971126 ns/op
BenchmarkPlus_Add_100000000-4   	       1	2084398768 ns/op
```

This is around `~50m` values per second.

Example using default `influx_stress` (which insert 100K series):

```
> select numSeries from "database"
name: database
time			numSeries
----			---------
1476213890000000000	99931
```

In terms of size, using a precision of `16`, each sketch takes up around `65KB`.

@benbjohnson @jwilder