@e-dard 2016-10-11
@pauldix you [asked](https://github.com/influxdata/influxdb/pull/7264#issuecomment-245345462) about the accuracy of the estimates for low cardinalities. Here are some other benchmarks I took that compare the actual cardinality to the estimation produced from the `HLL++`. `Add_100` means `100` items were added to the sketch. The benchmarks go from `100` to `100M`. For low cardinalities the implementation will revert to `LinearCounting`, which is an accurate algorithm for estimating low-cardinality data.

As you can see, the estimates are quite accurate. It took `2.62s` to add `100M` items. The time it takes to generate the count estimate is for all intents and purposes zero.

```
BenchmarkHLLPlus_Add_100-4         	  300000	      5971 ns/op
--- BENCH: BenchmarkHLLPlus_Add_100-4
	hll_test.go:301: True cardinality is 100
	hll_test.go:301: Estimated cardinality is 100
BenchmarkHLLPlus_Add_1000-4        	    3000	    369160 ns/op
--- BENCH: BenchmarkHLLPlus_Add_1000-4
	hll_test.go:301: True cardinality is 1000
	hll_test.go:301: Estimated cardinality is 1000
BenchmarkHLLPlus_Add_10000-4       	     100	  10164884 ns/op
--- BENCH: BenchmarkHLLPlus_Add_10000-4
	hll_test.go:301: True cardinality is 10000
	hll_test.go:301: Estimated cardinality is 10000
BenchmarkHLLPlus_Add_100000-4      	     500	   2468274 ns/op
--- BENCH: BenchmarkHLLPlus_Add_100000-4
	hll_test.go:301: True cardinality is 100000
	hll_test.go:301: Estimated cardinality is 100051
BenchmarkHLLPlus_Add_1000000-4     	      50	  23998780 ns/op
--- BENCH: BenchmarkHLLPlus_Add_1000000-4
	hll_test.go:301: True cardinality is 1000000
	hll_test.go:301: Estimated cardinality is 1000399
BenchmarkHLLPlus_Add_10000000-4    	       5	 255199483 ns/op
--- BENCH: BenchmarkHLLPlus_Add_10000000-4
	hll_test.go:301: True cardinality is 10000000
	hll_test.go:301: Estimated cardinality is 10011519
BenchmarkHLLPlus_Add_100000000-4   	       1	2619683671 ns/op
--- BENCH: BenchmarkHLLPlus_Add_100000000-4
	hll_test.go:301: True cardinality is 100000000
	hll_test.go:301: Estimated cardinality is 100008535
```